Tidy up the code so that it makes more sense
Create new functions btnLowerSameHigherClick, btnOkOrDoneClick called
by existing callback functions. The code goes to the new functions
instead of being in the callback functions, to make things more
organized and easier to understand. Also, startRound function tests if
currentRoundNumber > numberOfRounds and ends the game if it is, instead
of this test being done by a callback function. All these changes are
for the code to make more sense, putting things in the place where they
belong.